requirements.txt
# Image Processing & OCR
opencv-python==4.10.0.84
rapidocr_onnxruntime==1.4.3
av==14.0.1
scikit-image==0.25.0
thefuzz==0.22.1
# Adb Control
adbutils==2.8.0
uiautomator2==3.2.8
# Visual Debugger & Web Server
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
websockets==14.1
python-socketio==5.11.4
psutil==6.1.1
# GUI
gradio==5.21.0
# Configuration
pydantic==2.10.4
# Other
typing-extensions==4.12.2
python-dotenv==1.0.1
# ImportError: DLL load failed while importing onnxruntime_pybind11_state
onnxruntime==1.14.0
# A module that was compiled using NumPy 1.x cannot be run in NumPy 2.0.0 as it may crash.
numpy<2.0